+++ components/browser/toolbar/src/main/java/mozilla/components/browser/toolbar/display/DisplayToolbar.kt
@@ -139,6 +139,9 @@ internal class DisplayToolbar(
     // Margin between browser actions.
     internal var browserActionMargin = 0
 
+    // Set if progress bar should be at the top of the toolbar
+    internal var topProgressBar = false
+
     // Horizontal margin of URL Box (surrounding URL and page actions).
     internal var urlBoxMargin = 0
 
@@ -344,6 +347,7 @@ internal class DisplayToolbar(
     }
 
     // We layout the toolbar ourselves to avoid the overhead from using complex ViewGroup implementations
+    @Suppress("ComplexMethod")
     override fun onLayout(changed: Boolean, left: Int, top: Int, right: Int, bottom: Int) {
         // First we layout the navigation actions if there are any:
         //   +-------------+------------------------------------------------+
@@ -434,9 +438,11 @@ internal class DisplayToolbar(
         val urlLeft = navigationActionsWidth + iconWidth + urlBoxMargin
         urlView.layout(urlLeft, 0, urlLeft + urlView.measuredWidth, measuredHeight)
 
-        // The progress bar is going to be drawn at the bottom of the toolbar:
+        // The progress bar by default is going to be drawn at the bottom of the toolbar, top if defined:
 
-        progressView.layout(0, measuredHeight - progressView.measuredHeight, measuredWidth, measuredHeight)
+        progressView.layout(0, if (topProgressBar) 0 else measuredHeight - progressView.measuredHeight,
+            measuredWidth, if (topProgressBar) progressView.measuredHeight else measuredHeight
+        )
 
         // The URL box view (if exists) is positioned behind the icon, the url and page actions: